// src/app/models/str.ts
export class Structure {
    firstName: any;
    lastName: any;
    fatherName: any;
    motherName: any;
    streetName: any;
    locality: any;
    landmark: any;
    city: any;
    state: any;
    pincode: any;
    classApply: any;
    tAndc: any;
    userName: any;
    password: any
    cnfPassword: any
}